pipeline/requirements.txt
spacy>=3.7,<4.0
requests>=2.31
anthropic>=0.25
edge-tts>=6.1
ebooklib>=0.18
beautifulsoup4>=4.12
lxml>=5.0
python-dotenv>=1.0
